+++ upload/catalog/controller/extension/payment/ginger_afterpay.php
@@ -2,13 +2,15 @@
 require_once(DIR_SYSTEM.'library/ginger/autoload.php');
 
 use interfaces\GingerCustomerPersonalInformation;
+use interfaces\GingerTermsAndConditions;
 use banktwins\GingerBankPluginController;
 use components\GingerBankConfig;
 use interfaces\GingerTestAPIKey;
 
 class ControllerExtensionPaymentGingerAfterPay extends GingerBankPluginController implements
 GingerCustomerPersonalInformation,
-GingerTestAPIKey
+GingerTestAPIKey,
+GingerTermsAndConditions
 {
     /**
      * Payments module name