+++ upload/catalog/model/extension/payment/ginger_klarnadirectdebit.php
@@ -1,9 +1,0 @@
-<?php
-require_once(DIR_SYSTEM.'library/ginger/autoload.php');
-
-use banktwins\GingerBankModel;
-
-class ModelExtensionPaymentGingerKlarnaDirectDebit extends GingerBankModel
-{
-    protected $paymentName = 'ginger_klarnadirectdebit';
-}