+++ upload/catalog/model/extension/payment/ginger_sofort.php
@@ -1,9 +1,0 @@
-<?php
-require_once(DIR_SYSTEM.'library/ginger/autoload.php');
-
-use banktwins\GingerBankModel;
-
-class ModelExtensionPaymentGingerSofort extends GingerBankModel
-{
-    protected $paymentName = 'ginger_sofort';
-}